Moved RobotMath functions to Drive
Step 1 of appeasing Caleb's wrath
