+++ src/main/java/frc/robot/commands/drive/Drive.java
@@ -7,10 +7,7 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import frc.robot.OI;
 import frc.robot.RobotMap;
 import frc.robot.sensors.NavX;
-import frc.robot.sensors.PDP;
-import frc.robot.subsystems.Climber;
 import frc.robot.subsystems.Drivetrain;
-import frc.robot.util.RobotMath;
 
 public class Drive extends Command {
 
@@ -26,6 +23,30 @@ public class Drive extends Command {
         requires(Drivetrain.getInstance());
     }
 
+    public static double powerCurve(double x) {
+        // https://www.desmos.com/calculator/g07ukjj7bl
+        double curve = (0.5D * (Math.atan(Math.PI * (Math.abs(x) - 0.5D)))) + 0.5D;
+        return Math.copySign(curve, x);
+    }
+
+    public static void normalize(double[] array) {
+        double max = Math.abs(array[0]);
+        boolean normFlag = max > 1;
+
+        for (int i = 1; i < array.length; i++) {
+            if (Math.abs(array[i]) > max) {
+                max = Math.abs(array[i]);
+                normFlag = max > 1;
+            }
+        }
+
+        if (normFlag) {
+            for (int i = 0; i < array.length; i++) {
+                array[i] /= max;
+            }
+        }
+    }
+
     private void correctForPitch(double[] stickSpeeds) {//x-tip
         //if its over 45 there's no point in correcting it
         if (Math.abs(pitchDegrees) < 45) {
@@ -35,7 +56,7 @@ public class Drive extends Command {
             double correctionOffset = correctionFactor * (pitchDegrees - RobotMap.DRIVE.PITCH_THRESHOLD_DEGREES);
             stickSpeeds[0] += correctionOffset;
             stickSpeeds[1] += correctionOffset;
-            RobotMath.normalize(stickSpeeds);
+            normalize(stickSpeeds);
         }
     }
 
@@ -58,7 +79,7 @@ public class Drive extends Command {
 
     @Override
     protected void execute() {
-        double stickSpeeds[] = {RobotMath.powerCurve(OI.getInstance().getLeftY()), RobotMath.powerCurve(OI.getInstance().getRightY())};
+        double stickSpeeds[] = {powerCurve(OI.getInstance().getLeftY()), powerCurve(OI.getInstance().getRightY())};
         //double leftStick = RobotMath.powerCurve(OI.getInstance().getLeftY());
         //double rightStick = RobotMath.powerCurve(OI.getInstance().getRightY());
